% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench.R
\name{snv.bench}
\alias{snv.bench}
\title{snv.bench}
\usage{
snv.bench(nuseq_classifiers, truths, training_set = FALSE)
}
\arguments{
\item{nuseq_classifiers}{List of named data frames as output from nuseq-sc.}

\item{truths}{List of data.frames with snv information. Easiest to use the convenience function 'convert_to_pos' and add a fifth column ('labels') with snvs annotated as true_positive (1), false_positive (0) and unknown (NA).}

\item{training_set}{If TRUE, only snvs with ground truth data will be output. This is useful if this data is to be used for training a model to be applied to a different dataset with snv.cross.predict.}
}
\description{
This is a convenience function that takes the output files from nuseq-sc and combines them with a list of data frames with annotated snvs for input into other snv.predict functions. It will output a single snv.summary file concatenating all results and a vector of the low correlating classifiers recommended for subsequent analysis.
}
\examples{
snv.bench(nuseq_classifiers,truths,training_set=FALSE)
}
