% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv.cross.predict.R
\name{snv.cross.predict}
\alias{snv.cross.predict}
\title{snv.cross.predict}
\usage{
snv.cross.predict(train_data, test_data, classifiers = NULL,
  threshold = NULL)
}
\arguments{
\item{train_data}{Output from snv.bench. A data frame containing all derived snv classifiers with ground truths.}

\item{test_data}{Output from snv.bench. A data frame containing all derived snv classifiers (where training_set=FALSE, though not essential).}

\item{classifiers}{Output from snv.bench. A vector to keep low correlating classifiers only, all used if NULL.If applied to the training data, it will need to be used here.}

\item{threshold}{Defaults to 0.5. Values 0 to 1 can be input to tune predictions for specificity or sensitivity.}
}
\description{
This function makes a snv prediction model on one set of data and then applies it to another, typically one without the ground truth information required to build a cell type specific model.
}
\examples{
snv.cross.predict(train_data,test_data,classifiers,threshold)
}
