% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv.performance.R
\name{snv.performance}
\alias{snv.performance}
\title{snv.performance}
\usage{
snv.performance(snv_data, classifiers = NULL, test_index = NULL,
  threshold = NULL, model.method = "randomforest")
}
\arguments{
\item{snv_data}{Output from snv.benchmark. A data frame containing all derived snv classifiers with ground truth data appended.}

\item{classifiers}{Defaults to use all classifiers. We recommend providing a vector of the low correlating classifiers to use in building the prediction model. This is already output from snv.bench.}

\item{test_index}{A vector to determine which samples to train the model on (1) and then test (2). This can be user input to allow consistency between modelling methods if required.}

\item{threshold}{Defaults to 0.5. Values 0 to 1 can be input to tune predictions for specificity or sensitivity.}

\item{model.method}{Defaults to standard random forest, other options are "tuned_randomforest", "svm", and "tuned_svm".}
}
\description{
This function takes the output files from snv.benchmark, makes a snv prediction model and tests it on hold out data. The output is a confusion matrix, a set of snv predictions and a table for plotting a ROC curve (all applied to test data only, AUC is output in STDOUT).
}
\examples{
snv.performance(snv_data,classifiers,test_index=NULL,threshold=NULL,model.method="randomforest")
}
