% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv.predict.R
\name{snv.predict}
\alias{snv.predict}
\title{snv.predict}
\usage{
snv.predict(nuseq_classifiers, truths)
}
\arguments{
\item{nuseq_classifiers}{List of named data frames as output from nuseq-sc.}

\item{truths}{List of named data.frames with snv information. Easiest to use the convenience function 'convert_to_pos' and add a fifth column with snvs annotated as true_positive (1), false_positive (0) and unknown (NA).}
}
\description{
This function predicts snvs in single cell RNA-seq data by creating a model based on true positives provided by the user and false positives from synthetic spike-ins (currently only ERCCs are supported). It will create a directory in the working  directory called snv.predict.results containing a table of snv predictions.
}
\examples{
snv.predict(classifiers,truths)
}
